% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caoimhe_xlsx.R
\name{msigdb_pathways_to_list}
\alias{msigdb_pathways_to_list}
\title{msigdb pathways in a nice list}
\usage{
msigdb_pathways_to_list(msigdb_species = "Homo sapiens", msigdb_cat = "H")
}
\arguments{
\item{msigdb_species}{character one of msigdbr::msigdbr_show_species()}

\item{msigdb_cat}{character one of 'c("H", paste0("C", c(1:7)))',
see: gsea-msigdb.org/gsea/msigdb/collections.jsp}
}
\value{
msigdb_pathlist list object
}
\description{
msigdb pathways in a nice list
}
