% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caoimhe_xlsx.R
\name{msigdb_pway_xlsx}
\alias{msigdb_pway_xlsx}
\title{msigdb pathways genes parsed from XLSX tb}
\usage{
msigdb_pway_xlsx(fgseaRes, tb, padj, TAG)
}
\arguments{
\item{fgseaRes}{tibble of fgseaResults}

\item{tb}{tibble of XLSX results}

\item{padj}{numeric value for padj below which we return results}

\item{TAG}{character to tag output and name plots etc}
}
\value{
NULL, writes XLSX containing pathways as sheets, and genes in those significant pathway
}
\description{
msigdb pathways genes parsed from XLSX tb
}
