% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caoimhe_xlsx.R
\name{read_xlsx_glaucoma}
\alias{read_xlsx_glaucoma}
\title{Read data from XLSX in extdata}
\usage{
read_xlsx_glaucoma(XLSX)
}
\arguments{
\item{XLSX}{character filename}
}
\description{
Read data from XLSX in extdata
}
