% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analyses.R
\name{run_analyses}
\alias{run_analyses}
\title{Run fgsea analysis on XLSX}
\usage{
run_analyses()
}
\arguments{
\item{XLSX}{character filename}

\item{TAG}{character output naming scheme}
}
\description{
Run fgsea analysis on XLSX
}
