% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caoimhe_xlsx.R
\name{run_group_analysis}
\alias{run_group_analysis}
\title{Run fgsea analysis on grouped XLSX}
\usage{
run_group_analysis(XLSX, TAG)
}
\arguments{
\item{XLSX}{character 'all' or 'equal' to specify which data to load}

\item{TAG}{character to tag output and name plots etc}
}
\description{
Run fgsea analysis on grouped XLSX
}
