% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caoimhe_xlsx.R
\name{write_all_hallmarks}
\alias{write_all_hallmarks}
\title{msigdb pathways genes to XLSX}
\usage{
write_all_hallmarks(TAG = "All_Hallmark_pathways")
}
\arguments{
\item{TAG}{character to tag output}
}
\value{
NULL, writes XLSX containing pathways as sheets, and genes in those significant pathway
}
\description{
msigdb pathways genes to XLSX
}
